backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
httpx

openai>=1.0.0
pinecone>=7.0.0

torch
numpy
pillow

transformers
datasets
accelerate

scikit-learn
sentence-transformers
